WebDownloader/requirements.txt
amqp==2.6.1
aniso8601==8.0.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.9
flask-redis==0.4.0
Flask-RESTful==0.3.8
furl==2.1.0
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
kombu==4.6.11
MarkupSafe==1.1.1
marshmallow==3.7.1
matplotlib==3.3.1
networkx==2.5
numpy==1.19.2
orderedmultidict==1.0.1
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
validators==0.18.1
vine==1.3.0
Werkzeug==1.0.1
